Moved ExpandedCardsRepository.kt to repositories package
