+++ app/src/main/java/pl/llp/aircasting/ui/view/screens/settings/my_account/MyAccountController.kt
@@ -2,9 +2,9 @@ package pl.llp.aircasting.ui.view.screens.settings.my_account
 
 import android.content.Context
 import pl.llp.aircasting.data.local.LogoutService
-import pl.llp.aircasting.util.Settings
 import pl.llp.aircasting.ui.view.common.BaseController
 import pl.llp.aircasting.ui.view.screens.login.LoginActivity
+import pl.llp.aircasting.util.Settings
 
 
 class MyAccountController(
@@ -28,5 +28,4 @@ class MyAccountController(
             LoginActivity.startAfterSignOut(mContext)
         }
     }
-
 }
